requirements.txt
amqp==2.5.2
billiard==3.6.3.0
celery==4.4.2
click==7.1.2
Flask==1.1.2
gevent==20.5.0
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.8
MarkupSafe==1.1.1
python-dotenv==0.13.0
pytz==2020.1
redis==3.5.2
vine==1.3.0
Werkzeug==1.0.1
